+++ transport/nio/src/test/java/org/kaazing/gateway/transport/nio/TcpTransportTest.java
@@ -1,6 +1,6 @@
 /**
  * Copyright (c) 2007-2014 Kaazing Corporation. All rights reserved.
- * 
+ *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -8,9 +8,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package org.kaazing.gateway.transport.nio.internal;
+package org.kaazing.gateway.transport.nio;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -34,42 +34,46 @@ import org.junit.Test;
 import org.kaazing.gateway.resource.address.ResourceAddress;
 import org.kaazing.gateway.transport.BridgeAcceptor;
 import org.kaazing.gateway.transport.BridgeConnector;
+import org.kaazing.gateway.transport.nio.TcpTransport;
+import org.kaazing.gateway.transport.nio.internal.NioSocketAcceptor;
+import org.kaazing.gateway.transport.nio.internal.NioSocketConnector;
+import org.kaazing.gateway.transport.nio.internal.TcpExtensionFactory;
 
 public class TcpTransportTest {
-    
+
     private TcpTransport transport = new TcpTransport(new Properties());
 
     private ResourceAddress address;
-    
+
     @Before
     public void before() throws Exception {
         address = newResourceAddressFactory().newResourceAddress(new URI("tcp://localhost:8888"));
     }
-    
+
     @Test
     public void getAcceptorShouldReturnAcceptor() throws Exception {
         BridgeAcceptor acceptor = transport.getAcceptor();
         assertTrue(acceptor instanceof NioSocketAcceptor);
     }
-    
+
     @Test
     public void getAcceptorWithResourceAddressShouldReturnAcceptor() throws Exception {
         BridgeAcceptor acceptor = transport.getAcceptor(address);
         assertTrue(acceptor instanceof NioSocketAcceptor);
     }
-    
+
     @Test
     public void getConnectorShouldReturnAcceptor() throws Exception {
         BridgeConnector connector = transport.getConnector();
         assertTrue(connector instanceof NioSocketConnector);
     }
-    
+
     @Test
     public void getConnectorWithResourceAddressShouldReturnAcceptor() throws Exception {
         BridgeConnector connector = transport.getConnector(address);
         assertTrue(connector instanceof NioSocketConnector);
     }
-    
+
     @Test
     public void getExtensionsShouldReturnAvailableExtensions() {
         Collection<?> extensions = transport.getExtensions();
